Move shapes widgets out into a separate struct
